% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_renderer-textInput.R
\name{renderTextTableInput}
\alias{renderTextTableInput}
\title{Table output containing editable data}
\usage{
\code{renderTextTableInput(df)}

\code{renderTextTableInput(df, tableId = NA, tableClasses = "table table-bordered")}

\code{renderTextTableInput(df, tableId = "ui_table_id", tableClasses)}
}
\arguments{
\item{df}{data frame input to be rendered as an editable HTML table (text).}

\item{tableId}{single length character vector denoting the id of the corresponding 
HTML table element generated in the UI by \code{textTableInput}.}

\item{tableClasses}{character vector specifying the HTML classes to apply to the 
table generated using \code{df}. table denoting the id of the corresponding HTML 
table element generated in the UI by \code{textTableInput}.}
}
\description{

}
\details{
Note: Section documentation in progress

Shiny has a default Javascript dependency on the Bootstrap framework. Therefore the
 \code{tableClasses} vector should specify

To apply default Bootstrap CSS table styling, simply specify them here for inclusion into the genrated table. include them in this arg

can apply bootstrap themes to the page that can define the table output: https://shiny.rstudio.com/reference/shiny/latest/bootstrapPage.html
}
\references{
W3Schools \href{https://www.w3schools.com/bootstrap/bootstrap_tables.asp}{Bootstrap tables} resource.
}
\seealso{
\code{\link{textTableInput}}
}
