% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_ui-textInput.R
\name{textTableInput}
\alias{textTableInput}
\title{Create a text input table}
\usage{
\code{textTableInput(tableId)}
}
\arguments{
\item{tableId}{The id of the UI element in which the generated HTML table will be embedded.}
}
\value{
An HTML table consisting of editable cells that can be added to a UI definition.
}
\description{
Creates an HTML table entity containing a grid of editable input cells corresponding to the data frame object rendered by \code{renderTextTableInput} with the corresponding \code{tableId}.
}
\details{

}
\seealso{
\code{\link{renderTextTableInput}}
}
