% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_example_apps.R
\name{runAppExample}
\alias{runAppExample}
\title{Run demo shiny apps to test out shinyinputtables functionality}
\usage{
runAppExample(demoName)
}
\arguments{
\item{demoName}{Single length character vector denoting the name of the desired demo app to run.}
}
\description{
Run a selected shiny app demo that showcases particulars aspects
of the functionality provided by the \code{shinyinputtables} package extension for Shiny.
}
\details{
Currently the only supported demos are currently called "demo_01" and "demo_02".
Will describe each available demo in a separate section here
}
